Moved top bar to separated component.
