+++ package.json
@@ -14,7 +14,6 @@
   "dependencies": {
     "@intlify/unplugin-vue-i18n": "^0.12.3",
     "dayjs": "^1.11.8",
-    "filename-reserved-regex": "^3.0.0",
     "pinia": "^2.0.36",
     "vue": "^3.3.2",
     "vue-i18n": "^v9.3.0-beta.21",
@@ -25,13 +24,14 @@
   "devDependencies": {
     "@mdi/font": "^7.2.96",
     "@rushstack/eslint-patch": "^1.2.0",
-    "@tauri-apps/api": "^1.4.0",
-    "@tauri-apps/cli": "^1.4.0",
+    "@tauri-apps/api": "^1.5.3",
+    "@tauri-apps/cli": "^1.5.9",
     "@vitejs/plugin-vue": "^4.6.2",
     "@vue/eslint-config-prettier": "^7.1.0",
     "@vue/test-utils": "^2.3.2",
     "eslint": "^8.39.0",
     "eslint-plugin-vue": "^9.11.0",
+    "filename-reserved-regex": "^3.0.0",
     "jsdom": "^22.1.0",
     "prettier": "^2.8.8",
     "vite": "^4.5.2",
