+++ src-tauri/tauri.conf.json
@@ -121,8 +121,7 @@
         "height": 800,
         "minWidth": 1024,
         "minHeight": 768,
-        "decorations": false,
-        "fileDropEnabled": false
+        "decorations": false
       }
     ]
   }
